Module Luv.Passwd

type user = {

username : string;

uid : Unsigned.ulong;

gid : Unsigned.ulong;

shell : string option;

homedir : string;

}

Binds uv_passwd_t.

val get_passwd : ?uid:Unsigned.ulong -> unit -> (user, Error.t) Stdlib.result

Gets passwd entry for the current user or the user with the given uid.

Binds uv_os_get_passwd. See getpwuid_r(3p).

Requires libuv 1.9.0.

Feature check: Luv.Require.(has os_get_passwd)

The ?uid argument requires Luv 0.5.13 and libuv 1.45.0.

Feature check: Luv.Require.(has os_get_passwd_uid)

type group = {

groupname : string;

gid : Unsigned.ulong;

members : string list;

}

Binds uv_group_t.

val get_group : Unsigned.ulong -> (group, Error.t) Stdlib.result

Gets the entry for the group with the given gid.

Binds uv_os_get_group.

Requires Luv 0.5.13 and libuv 1.45.0.

Feature check: Luv.Require.(has os_get_group)