Module Luv.Request

Requests.

See Requests in the user guide and uv_req_t — Base request in libuv.

Requests are objects libuv uses to track asynchronous operations, and sometimes to communicate their results. For the most part Luv handles requests automatically.

Some request kinds support cancelation, and Luv provides a common function Luv.Request.cancel for them.

Apart from that, this module would be only an internal convenience for the implementation of Luv.

The full list of exposed concrete request types:

Luv.File.Request.t

Luv.DNS.Addr_info.Request.t

Luv.DNS.Name_info.Request.t

Luv.Thread_pool.Request.t

type 'kind t = 'kind Luv_c_types.Request.t Ctypes.ptr

Binds uv_req_t.

val cancel : [< `File | `Addr_info | `Name_info | `Random | `Thread_pool ] t -> (unit, Error.t) Stdlib.result

Tries to cancel a pending request.

Binds uv_cancel.