Module Luv.Version

Version information for the vendored libuv.

See Version-checking macros and functions in libuv.

Luv currently vendors libuv 1.30.1.

val string : unit -> string

Returns the libuv version as a string. See uv_version_string.

val major : int

libuv major version number. See UV_VERSION_MAJOR.

val minor : int

libuv minor version number. See UV_VERSION_MINOR.

val patch : int

libuv patch number. See UV_VERSION_PATCH.

val is_release : bool

true if the libuv version is a release, and false if it is a development version. This will almost always be true for Luv releases. See UV_VERSION_IS_RELEASE.

val suffix : string

libuv version suffix for development releases. See UV_VERSION_SUFFIX.

val hex : int

libuv version packed into a single integer. See UV_VERSION_HEX.

val version : unit -> int

Returns Luv.Version.hex. See uv_version.