Module Luv__.DNS

module Addr_info : sig ... end

Binds struct addrinfo and request objects for Luv.DNS.getaddrinfo.

module Name_info : sig ... end

Optional flags and request objects for use with Luv.DNS.getnameinfo.

val getaddrinfo : ?⁠loop:Luv.Loop.t -> ?⁠request:Addr_info.Request.t -> ?⁠family:Luv.Sockaddr.Address_family.t -> ?⁠socktype:Luv.Sockaddr.Socket_type.t -> ?⁠protocol:int -> ?⁠flags:Addr_info.Flag.t list -> ?⁠node:string -> ?⁠service:string -> unit -> ((Addr_info.t list, Luv.Error.t) Result.result -> unit) -> unit

Retrieves addresses.

Binds uv_getaddrinfo. See getaddrinfo(3).

uv_getaddrinfo and getaddrinfo(3) take optional hints in fields of an argument of type struct addrinfo. Luv.DNS.getaddrinfo instead has several optional arguments, each named after one of the fields of hints.

val getnameinfo : ?⁠loop:Luv.Loop.t -> ?⁠request:Name_info.Request.t -> ?⁠flags:Name_info.Flag.t list -> Luv.Sockaddr.t -> ((string * string, Luv.Error.t) Result.result -> unit) -> unit

Retrieves host names.

Binds uv_getnameinfo. See getnameinfo(3).