Module Luv__Error

Error handling.

See Error handling in the user guide and Error handling in libuv.

type t = [

| `E2BIG

| `EACCES

| `EADDRINUSE

| `EADDRNOTAVAIL

| `EAFNOSUPPORT

| `EAGAIN

| `EAI_ADDRFAMILY

| `EAI_AGAIN

| `EAI_BADFLAGS

| `EAI_BADHINTS

| `EAI_CANCELED

| `EAI_FAIL

| `EAI_FAMILY

| `EAI_MEMORY

| `EAI_NODATA

| `EAI_NONAME

| `EAI_OVERFLOW

| `EAI_PROTOCOL

| `EAI_SERVICE

| `EAI_SOCKTYPE

| `EALREADY

| `EBADF

| `EBUSY

| `ECANCELED

| `ECONNABORTED

| `ECONNREFUSED

| `ECONNRESET

| `EDESTADDRREQ

| `EEXIST

| `EFAULT

| `EFBIG

| `EFTYPE

| `EHOSTUNREACH

| `EILSEQ

| `EINTR

| `EINVAL

| `EIO

| `EISCONN

| `EISDIR

| `ELOOP

| `EMFILE

| `EMSGSIZE

| `ENAMETOOLONG

| `ENETDOWN

| `ENETUNREACH

| `ENFILE

| `ENOBUFS

| `ENODEV

| `ENOENT

| `ENOMEM

| `ENONET

| `ENOPROTOOPT

| `ENOSPC

| `ENOSYS

| `ENOTCONN

| `ENOTDIR

| `ENOTEMPTY

| `ENOTSOCK

| `ENOTSUP

| `ENOTTY

| `EPERM

| `EPIPE

| `EPROTO

| `EPROTONOSUPPORT

| `EPROTOTYPE

| `ERANGE

| `EROFS

| `ESHUTDOWN

| `ESPIPE

| `ESRCH

| `ETIMEDOUT

| `ETXTBSY

| `EXDEV

| `UNKNOWN

| `EOF

| `ENXIO

| `EMLINK

]

Error codes returned by libuv functions.

Binds libuv error codes, which resemble Unix error codes.

val strerror : t -> string

Returns the error message corresponding to the given error code.

Binds uv_strerror_r.

val err_name : t -> string

Returns the name of the given error code.

Binds uv_err_name_r.

val translate_sys_error : int -> t

Converts a system error code to a libuv error code.

Binds uv_translate_sys_error.

val set_on_unhandled_exception : (exn -> unit) -> unit

If user code terminates a callback by raising an exception, the exception cannot be allowed to go up the call stack, because the callback was called by libuv (rather than OCaml code), and the exception would disrupt libuv book-keeping. Luv instead passes the exception to a global Luv exception handler. Luv.Error.set_on_unhandled_exception f replaces this exception handler with f.

For example, in

Luv.Error.set_on_unhandled_exception f;
Luv.File.mkdir "foo" (fun _ -> raise Exit);

the exception Exit is passed to f when mkdir calls its callback.

It is recommended to avoid letting exceptions escape from callbacks in this way.

The default behavior, if Luv.Error.set_on_unhandled_exception is never called, is for Luv to print the exception to STDERR and exit the process with exit code 2.

It is recommended not to call Luv.Error.set_on_unhandled_exception from libraries based on Luv, but, instead, to leave the decision on how to handle exceptions up to the final application.