Module Luv__UDP.Connected

Connected UDP sockets.

val connect : t -> Luv.Sockaddr.t -> (unit, Luv.Error.t) Result.result

Assigns a peer address to the given socket.

Binds uv_udp_connect.

val disconnect : t -> (unit, Luv.Error.t) Result.result

Removes the peer address assigned to the given socket.

Binds uv_udp_connect with NULL argument.

val getpeername : t -> (Luv.Sockaddr.t, Luv.Error.t) Result.result

Retrieves the peer address assigned to the given socket.

Binds uv_udp_getpeername.

val send : t -> Luv.Buffer.t list -> ((unit, Luv.Error.t) Result.result -> unit) -> unit

Like Luv.UDP.send, but the remote address used is the peer address assigned to the socket.

Binds uv_udp_send.

val try_send : t -> Luv.Buffer.t list -> (unit, Luv.Error.t) Result.result

Like Luv.UDP.try_send, but the remote address used is the peer address assigned to the socket.

Binds uv_udp_try_send.