Module Luv.Idle

Per-iteration callback.

See Example in the user guide and uv_idle_t — Idle handle in libuv.

type t = [ `Idle ] Handle.t

Binds uv_idle_t.

Note that values of this type can be passed to functions in Luv.Handle, in addition to the functions in this module. In particular, see Luv.Handle.close.

val init : ?⁠loop:Loop.t -> unit -> (t, Error.t) Result.result

Allocates and initializes an idle handle.

Binds uv_idle_init.

The handle should be cleaned up with Luv.Handle.close when no longer needed.

val start : t -> (unit -> unit) -> (unit, Error.t) Result.result

Starts the handle with the given callback.

Binds uv_idle_start.

val stop : t -> (unit, Error.t) Result.result

Stops the handle.

Binds uv_idle_stop.